+++ src/main/java/com/atlauncher/gui/tabs/ToolsTab.java
@@ -19,6 +19,7 @@ package com.atlauncher.gui.tabs;
 
 import com.atlauncher.data.Language;
 import com.atlauncher.gui.components.BlankToolPanel;
+import com.atlauncher.gui.components.DownloadClearerToolPanel;
 import com.atlauncher.gui.components.LogClearerToolPanel;
 import com.atlauncher.gui.components.NetworkCheckerToolPanel;
 import com.atlauncher.gui.components.ServerCheckerToolPanel;
@@ -40,8 +41,8 @@ public class ToolsTab extends JPanel implements Tab {
 
         mainPanel.add(new NetworkCheckerToolPanel());
         mainPanel.add(new ServerCheckerToolPanel());
+        mainPanel.add(new DownloadClearerToolPanel());
         mainPanel.add(new LogClearerToolPanel());
-        mainPanel.add(new BlankToolPanel());
         mainPanel.add(new BlankToolPanel());
         mainPanel.add(new BlankToolPanel());