+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Changelog
 
 - Update code to use Java 7
 - Mac OSX app updated to support Java 7 and no longer use Java 6. Users must redownload the app
+- Added in Download Clearer Tool to remove all your downloads from downloading packs to clear space if needed